requirements.txt
kafka-python
attrs
